txncorrectnesstest: move more tests to their own files
