+++ pkg/kv/kvclient/kvcoord/txncorrectnesstest/phantom_delete_anomaly_test.go
@@ -1,0 +1,39 @@
+// Copyright 2025 The Cockroach Authors.
+//
+// Use of this software is governed by the CockroachDB Software License
+// included in the /LICENSE file.
+
+package txncorrectnesstest
+
+import (
+	"testing"
+
+	"github.com/cockroachdb/cockroach/pkg/util/leaktest"
+	"github.com/cockroachdb/cockroach/pkg/util/log"
+	"github.com/cockroachdb/errors"
+)
+
+// TestTxnDBPhantomDeleteAnomaly verifies that neither SI nor SSI
+// isolation are subject to the phantom deletion anomaly; this is
+// similar to phantom reads, but verifies the delete range
+// functionality causes read/write conflicts.
+//
+// Phantom deletes would typically fail with a history such as:
+//
+//	R2(B) DR1(A-C) I2(B) C2 SC1(A-C) W1(D,A+B) C1
+func TestTxnDBPhantomDeleteAnomaly(t *testing.T) {
+	defer leaktest.AfterTest(t)()
+	defer log.Scope(t).Close(t)
+	txn1 := "DR(A-C) SC(A-C) W(D,A+B) C"
+	txn2 := "R(B) I(B) C"
+	verify := &verifier{
+		history: "R(D)",
+		checkFn: func(env map[string]int64) error {
+			if env["D"] != 0 {
+				return errors.Errorf("expected delete range to yield an empty scan of same range, sum=%d", env["D"])
+			}
+			return nil
+		},
+	}
+	checkConcurrency("phantom delete", serializableAndSnapshot, []string{txn1, txn2}, verify, t)
+}
