+++ pkg/kv/kvclient/kvcoord/txncorrectnesstest/write_skew_anomaly_test.go
@@ -1,0 +1,41 @@
+// Copyright 2025 The Cockroach Authors.
+//
+// Use of this software is governed by the CockroachDB Software License
+// included in the /LICENSE file.
+
+package txncorrectnesstest
+
+import (
+	"testing"
+
+	"github.com/cockroachdb/cockroach/pkg/kv/kvserver/concurrency/isolation"
+	"github.com/cockroachdb/cockroach/pkg/util/leaktest"
+	"github.com/cockroachdb/cockroach/pkg/util/log"
+)
+
+// TestTxnDBWriteSkewAnomaly verifies that SI suffers from the write
+// skew anomaly but not SSI. The write skew anomaly is a condition which
+// illustrates that snapshot isolation is not serializable in practice.
+//
+// With write skew, two transactions both read values from A and B
+// respectively, but each writes to either A or B only. Thus there are
+// no write/write conflicts but a cycle of dependencies which result in
+// "skew". Only serializable isolation prevents this anomaly.
+//
+// Write skew would typically fail with a history such as:
+//
+//	SC1(A-C) SC2(A-C) W1(A,A+B+1) C1 W2(B,A+B+1) C2
+//
+// In the test below, each txn reads A and B and increments one by 1.
+// The read values and increment are then summed and written either to
+// A or B. If we have serializable isolation, then the final value of
+// A + B must be equal to 3 (the first txn sets A or B to 1, the
+// second sets the other value to 2, so the total should be
+// 3). Snapshot isolation, however, may not notice any conflict (see
+// history above) and may set A=1, B=1.
+func TestTxnDBWriteSkewAnomaly(t *testing.T) {
+	defer leaktest.AfterTest(t)()
+	defer log.Scope(t).Close(t)
+	runWriteSkewTest(t, isolation.Serializable)
+	runWriteSkewTest(t, isolation.Snapshot)
+}
